+++ bin/displayAccuMesh.cpp
@@ -14,21 +14,15 @@
 #include <DGtal/images/ImageContainerBySTLMap.h>
 #include <DGtal/kernel/sets/DigitalSetFromMap.h>
 
-
-#include <boost/program_options/options_description.hpp>
-#include <boost/program_options/parsers.hpp>
-#include <boost/program_options/variables_map.hpp>
+#include "CLI11.hpp"
 
 
 #include "AccumulatorHelper.h"
 #include "NormalAccumulator.h"
 
 
-
-
 using namespace std;
 using namespace DGtal;
-namespace po = boost::program_options;
 
 
 
@@ -36,9 +30,6 @@ namespace po = boost::program_options;
 typedef ImageContainerBySTLVector<Z3i::Domain, unsigned int> Image3D;
 typedef ImageContainerBySTLVector<Z3i::Domain, unsigned char> Image3DChar;
 typedef ImageContainerBySTLVector<Z3i::Domain,  Z3i::RealPoint> ImageVector;
-
-
-
 
 
 typedef ImageContainerBySTLMap<Z3i::Domain,double> DistanceImage;
@@ -92,64 +83,60 @@ applyMedian(const TImage &anImage, unsigned int size){
  */
 int main(int argc, char *const *argv)
 {
-  po::options_description general_opt("Allowed options are: ");
-  general_opt.add_options()
-    ("help,h", "display this message")
-    ("input,i", po::value<std::string>(), "input mesh.")
-    ("output,o", po::value<std::string>(), "output mesh.")
-    ("minAccumulation,m", po::value<double>()->default_value(0.0), "min accumulation or confidence to consider display.")
-    ("cutoff", po::value<double>()->default_value(10000.0), "max cut off display.")
-    ("estimRadiusType", po::value<std::string>()->default_value("mean"), "set the type of the radius estimation (mean, min, median or max).")
-    ("displayFromConfidence", "flag to display value from confident image insted default accumulation.")
-    ("invertNormal", "invert input normal.")
-    ("filterFacesNoAcc", "filter faces with no accumulation.")
-    ("importNormals", po::value<std::string>(), "Use imported normals instead the one computed from the mesh faces.")
-    ("filterFacesMaxRadius,R", po::value<double>()->default_value(100000.0), "filter faces with radius greater than a threshold.")
-    ("filterFacesMinRadius", po::value<double>()->default_value(0.0), "filter faces with radius less than a threshold.")
-    ("segmentColor", "filter and display mesh faces in color (to be used with --segmentTh) .")
-    ("segmentThreshold", po::value<double>()->default_value(0.0), "set the segmentation thershold.")
-    ("radius,r", po::value<double>()->default_value(5), "maximal radius of accumulation analysis.");
-  
-  
-  bool parseOK = true;
-  po::variables_map vm;
-  
-  
-  try {
-    po::store(po::parse_command_line(argc, argv, general_opt), vm);
-  } catch (const std::exception &ex) {
-    trace.info() << "Error checking program options: " << ex.what() << std::endl;
-    parseOK = false;
-  }
 
-  po::notify(vm);
+  string inputFile;
+  string outputFile {"result.off"};
+  double minAccumulation {0.0};
+  double radius {5.0};
+  double cutoff {10000.0};
+  std::string estimRadiusType {"mean"};
+  bool invertNormal {false};
+  bool useConfident {false};
+  bool filterFacesNoAcc {false};
+  string importNormals {""};
+  double maxRadiusFilter {100000.0};
 
-  if (vm.count("help") || argc <= 1 || !parseOK || !vm.count("input")) {
-    trace.info() << "Display on mesh accumulation values from normal vector" << std::endl <<
-                 "Options: " << std::endl
-                 << general_opt << "\n";
-    return 0;
-  }
-  
-  double radius = vm["radius"].as<double>();
-  string inputFile = vm["input"].as<std::string>();
-  string outputFile = vm["output"].as<std::string>();
-  bool invertNormal = vm.count("invertNormal");
-  std::string estimRadiusType = vm["estimRadiusType"].as<std::string>();
-  bool useConfident = vm.count("displayFromConfidence");
-  double minAccumulation = vm["minAccumulation"].as<double>();
-  double maxRadiusFilter = vm["filterFacesMaxRadius"].as<double>();
-  double minRadiusFilter = vm["filterFacesMinRadius"].as<double>();
-  bool segmentColor = vm.count("segmentColor");
-  double segmentThreshold = vm["segmentThreshold"].as<double>();
-  double cutoff = vm["cutoff"].as<double>();
-  bool importNormals = vm.count("importNormals");
+  double minRadiusFilter {0.0};
+  bool segmentColor {false};
+  double segmentThreshold {0.0};
   
   
   DGtal::Color colorInf = DGtal::Color::Blue;
   DGtal::Color colorSup = DGtal::Color::Yellow;
   DGtal::Color colorNoAcc = DGtal::Color(100, 100, 100, 30);
   
+  
+  
+  // parse command line using CLI ----------------------------------------------
+  // parse command line using CLI ----------------------------------------------
+  CLI::App app;
+
+  app.description("Display on mesh accumulation values from normal vector ");
+  app.add_option("-i,--input,1", inputFile, "input mesh." )
+      ->required()
+      ->check(CLI::ExistingFile);
+  app.add_option("--output,-o,2", outputFile, "output mesh.", true);
+  app.add_option("--minAccumulation,-m", minAccumulation, "min accumulation or confidence to consider display.");
+  app.add_option("--radius,-r", radius, "maximal radius of accumulation analysis.", true);
+  app.add_option("--cutoff", radius, "max cut off display.", true);
+
+  app.add_option("--radiusEstimator,-e", estimRadiusType,  "set the type of the radius estimation (mean (default), min, median or max).", true)
+  -> check(CLI::IsMember({"max", "min", "mean", "median"}));
+  app.add_option("--importNormals", importNormals,"Use imported normals instead the one computed from the mesh faces." );
+  auto maxRadiusFilterOpt = app.add_option("--filterFacesMaxRadius,-R",maxRadiusFilter, "filter faces with radius greater than a threshold.", true);
+  auto minRadiusFilterOpt = app.add_option("--filterFacesMinRadius", minRadiusFilter,"filter faces with radius less than a threshold.", true);
+  app.add_option("--segmentThreshold", segmentThreshold,"set the segmentation thershold.", true);
+  app.add_flag("--invertNormal", invertNormal, "invert input normal.");
+  app.add_flag("--displayFromConfidence", useConfident, "flag to display value from confident image insted default accumulation.");
+  app.add_flag("--filterFacesNoAcc", filterFacesNoAcc, "filter faces with no accumulation." );
+  app.add_flag("--segmentColor", segmentColor, "filter and display mesh faces in color (to be used with --segmentTh).");
+  
+  app.get_formatter()->column_width(40);
+  CLI11_PARSE(app, argc, argv);
+  // END parse command line using CLI ----------------------------------------------
+
+    
+   
   // 1) Reading input mesh
   Mesh<Z3i::RealPoint> aMesh(true);
   aMesh << inputFile;
@@ -161,9 +148,9 @@ int main(int argc, char *const *argv)
   trace.info()<< "Computing accumulation image .... on domain: " << bb;
 
   NormalAccumulator normAcc (radius, estimRadiusType);
-  if(importNormals){
+  if(importNormals != ""){
     std::vector<DGtal::Z3i::RealPoint> vectorNorm =
-      PointListReader<DGtal::Z3i::RealPoint>::getPointsFromFile(vm["importNormals"].as<std::string>());
+      PointListReader<DGtal::Z3i::RealPoint>::getPointsFromFile(importNormals);
     normAcc.initFromMeshAndNormals(aMesh, vectorNorm, invertNormal);
     
   }else{
@@ -210,9 +197,9 @@ int main(int argc, char *const *argv)
       else  {
         aMesh.setFaceColor(i, colorSup );
       }
-    }else  if((val==0 && vm.count("filterFacesNoAcc"))||
-              (vm.count("filterFacesMinRadius") && val  < minRadiusFilter)||
-              (vm.count("filterFacesMaxRadius") && val > maxRadiusFilter)){
+    }else  if((val==0 && filterFacesNoAcc)||
+              (*minRadiusFilterOpt && val  < minRadiusFilter)||
+              (*maxRadiusFilterOpt && val > maxRadiusFilter)){
       faceToBeRemoved.push_back(i);
     }
     if(!segmentColor){      
@@ -223,22 +210,11 @@ int main(int argc, char *const *argv)
   
   DGtal::trace.info() << "[done]" << std::endl;
   trace.info()<< "Val max dist accumulation: " << maxVal << std::endl;
-  if(!segmentColor && (vm.count("filterFacesNoAcc") ||vm.count("filterFacesMinRadius")
-                       || vm.count("filterFacesMaxRadius")) ){
+  if(!segmentColor && (filterFacesNoAcc || *minRadiusFilterOpt
+                       || *maxRadiusFilterOpt) ){
     aMesh.removeFaces(faceToBeRemoved);
   }
   
   aMesh >> outputFile  ;
   return 0;
 }
-
-
-
-
-
-
-
-
-
-
-