+++ Changelog.md
@@ -1,7 +1,10 @@
 
 # Version 1.1
+  - Replace boost prog option with Cli11
+   (Bertrand [#24](Kerautret https://github.com/kerautret/CDCVAM/pull/24))
   - Fix compilation issues (missing std::)
    (Bertrand [#23](Kerautret https://github.com/kerautret/CDCVAM/pull/23))
+
 
 # Version 1.0
  initial version